refactored download button separately from treeNode component
